requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Google Calendar API
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# HTTP client for RAG sync
httpx==0.27.2

# Environment variables
python-dotenv==1.0.1

# Logging (built-in, but explicit for clarity)
# Python 3.11 includes logging module
